Modified mouse look to reorient the model instead of the physics object. Removed unnecessary delta tracking in mouselook.
`Input.GetAxisRaw` already does delta tracking internally, so we were taking deltas of deltas.
The newX/oldX/diffX stuff was basically just doing a bunch of work to just go and undo it so it's gone too. (I think it's also not quite right but I didn't sit down to look at it closely.

The bird is officially fun to fly going zoom fast :)